+++ settingrepo-impl/src/main/java/com/dwarfeng/settingrepo/impl/service/ImageListNodeMaintainServiceImpl.java
@@ -168,7 +168,7 @@ public class ImageListNodeMaintainServiceImpl implements ImageListNodeMaintainSe
     @BehaviorAnalyse
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public List<StringIdKey> batchInsertIfNotExists(List<ImageListNode> entities) throws ServiceException {
+    public List<StringIdKey> batchInsertIfNotExists(@SkipRecord List<ImageListNode> entities) throws ServiceException {
         return crudService.batchInsertIfNotExists(entities);
     }