+++ settingrepo-impl/src/main/java/com/dwarfeng/settingrepo/impl/service/ImageNodeMaintainServiceImpl.java
@@ -168,7 +168,7 @@ public class ImageNodeMaintainServiceImpl implements ImageNodeMaintainService {
     @BehaviorAnalyse
     @SkipRecord
     @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
-    public List<StringIdKey> batchInsertIfNotExists(List<ImageNode> entities) throws ServiceException {
+    public List<StringIdKey> batchInsertIfNotExists(@SkipRecord List<ImageNode> entities) throws ServiceException {
         return crudService.batchInsertIfNotExists(entities);
     }